Renamed TaskManager._issue_command to BrowserManagerHandle.execute_command_sequence
